refactor: factor out solver unit tests
