+++ cadical/src/lib.rs
@@ -797,178 +797,18 @@ impl fmt::Display for Limit {
 
 #[cfg(test)]
 mod test {
-    use std::{
-        sync::{Arc, Mutex},
-        thread,
-    };
-
     use super::{CaDiCaL, Config, Limit};
     use rustsat::{
         lit,
-        solvers::{
-            ControlSignal, FreezeVar, Learn, Solve, SolverResult, SolverState, StateError,
-            Terminate,
-        },
+        solvers::{Solve, SolverState, StateError},
         types::TernaryVal,
         var,
     };
 
-    #[test]
-    fn build_destroy() {
-        let _solver = CaDiCaL::default();
-    }
-
-    #[test]
-    fn build_two() {
-        let _solver1 = CaDiCaL::default();
-        let _solver2 = CaDiCaL::default();
-    }
-
-    #[test]
-    fn tiny_instance_sat() {
-        let mut solver = CaDiCaL::default();
-        solver.add_binary(lit![0], !lit![1]).unwrap();
-        solver.add_binary(lit![1], !lit![2]).unwrap();
-        let ret = solver.solve();
-        match ret {
-            Err(e) => panic!("got error when solving: {}", e),
-            Ok(res) => assert_eq!(res, SolverResult::Sat),
-        }
-    }
-
-    #[test]
-    fn tiny_instance_unsat() {
-        let mut solver = CaDiCaL::default();
-        solver.add_unit(!lit![0]).unwrap();
-        solver.add_binary(lit![0], !lit![1]).unwrap();
-        solver.add_binary(lit![1], !lit![2]).unwrap();
-        solver.add_unit(lit![2]).unwrap();
-        let ret = solver.solve();
-        match ret {
-            Err(e) => panic!("got error when solving: {}", e),
-            Ok(res) => assert_eq!(res, SolverResult::Unsat),
-        }
-    }
-
-    #[test]
-    fn tiny_instance_multithreaded_sat() {
-        let mutex_solver = Arc::new(Mutex::new(CaDiCaL::default()));
-        let ret = {
-            let mut solver = mutex_solver.lock().unwrap();
-            solver.add_binary(lit![0], !lit![1]).unwrap();
-            solver.add_binary(lit![1], !lit![2]).unwrap();
-            solver.solve()
-        };
-        match ret {
-            Err(e) => panic!("got error when solving: {}", e),
-            Ok(res) => assert_eq!(res, SolverResult::Sat),
-        }
-
-        // Now in another thread
-        let s = mutex_solver.clone();
-        let ret = thread::spawn(move || {
-            let mut solver = s.lock().unwrap();
-            solver.solve()
-        })
-        .join()
-        .unwrap();
-        match ret {
-            Err(e) => panic!("got error when solving: {}", e),
-            Ok(res) => assert_eq!(res, SolverResult::Sat),
-        }
-
-        // Now add to solver in other thread
-        let s = mutex_solver.clone();
-        let ret = thread::spawn(move || {
-            let mut solver = s.lock().unwrap();
-            solver.add_unit(!lit![0]).unwrap();
-            solver.add_unit(lit![2]).unwrap();
-            solver.solve()
-        })
-        .join()
-        .unwrap();
-        match ret {
-            Err(e) => panic!("got error when solving: {}", e),
-            Ok(res) => assert_eq!(res, SolverResult::Unsat),
-        }
-
-        // Finally, back in the main thread
-        let ret = {
-            let mut solver = mutex_solver.lock().unwrap();
-            solver.add_binary(lit![0], !lit![1]).unwrap();
-            solver.add_binary(lit![1], !lit![2]).unwrap();
-            solver.solve()
-        };
-        match ret {
-            Err(e) => panic!("got error when solving: {}", e),
-            Ok(res) => assert_eq!(res, SolverResult::Unsat),
-        }
-    }
-
-    #[test]
-    fn termination_callback() {
-        let mut solver = CaDiCaL::default();
-        solver.add_binary(lit![0], !lit![1]).unwrap();
-        solver.add_binary(lit![1], !lit![2]).unwrap();
-        solver.add_binary(lit![2], !lit![3]).unwrap();
-        solver.add_binary(lit![3], !lit![4]).unwrap();
-        solver.add_binary(lit![4], !lit![5]).unwrap();
-        solver.add_binary(lit![5], !lit![6]).unwrap();
-        solver.add_binary(lit![6], !lit![7]).unwrap();
-        solver.add_binary(lit![7], !lit![8]).unwrap();
-        solver.add_binary(lit![8], !lit![9]).unwrap();
-
-        solver.attach_terminator(|| ControlSignal::Terminate);
-
-        let ret = solver.solve();
-
-        match ret {
-            Err(e) => panic!("got error when solving: {}", e),
-            Ok(res) => assert_eq!(res, SolverResult::Interrupted),
-        }
-
-        // Note: since IPASIR doesn't specify _when_ the terminator callback needs
-        // to be called, there is no guarantee that the callback is actually
-        // called during solving. This might cause this test to fail with some solvers.
-    }
-
-    #[test]
-    fn learner_callback() {
-        let mut solver = CaDiCaL::default();
-        solver.add_binary(lit![0], !lit![1]).unwrap();
-        solver.add_binary(lit![1], !lit![2]).unwrap();
-        solver.add_binary(lit![2], !lit![3]).unwrap();
-        solver.add_binary(lit![3], !lit![4]).unwrap();
-        solver.add_binary(lit![4], !lit![5]).unwrap();
-        solver.add_binary(lit![5], !lit![6]).unwrap();
-        solver.add_binary(lit![6], !lit![7]).unwrap();
-        solver.add_binary(lit![7], !lit![8]).unwrap();
-        solver.add_binary(lit![8], !lit![9]).unwrap();
-        solver.add_unit(lit![9]).unwrap();
-        solver.add_unit(!lit![0]).unwrap();
-
-        let mut cl_len = 0;
-
-        solver.attach_learner(
-            |clause| {
-                cl_len = clause.len();
-            },
-            10,
-        );
-
-        let ret = solver.solve();
-
-        drop(solver);
-
-        // Note: it is hard to create a testing instance on which clause learning
-        // actually happens and therefore it is not actually tested if the
-        // callback is called
-
-        match ret {
-            Err(e) => panic!("got error when solving: {}", e),
-            Ok(res) => assert_eq!(res, SolverResult::Unsat),
-        }
-    }
+    rustsat_solvertests::basic_unittests!(CaDiCaL);
+    rustsat_solvertests::termination_unittests!(CaDiCaL);
+    rustsat_solvertests::learner_unittests!(CaDiCaL);
+    rustsat_solvertests::freezing_unittests!(CaDiCaL);
 
     #[test]
     fn configure() {
@@ -1018,20 +858,6 @@ mod test {
         assert_eq!(solver.get_irredundant(), 9);
         assert_eq!(solver.get_redundant(), 0);
         assert_eq!(solver.current_lit_val(lit![0]), TernaryVal::DontCare);
-    }
-
-    #[test]
-    fn freezing() {
-        let mut solver = CaDiCaL::default();
-        solver.add_binary(lit![0], !lit![1]).unwrap();
-
-        solver.freeze_var(var![0]).unwrap();
-
-        assert!(solver.is_frozen(var![0]).unwrap());
-
-        solver.melt_var(var![0]).unwrap();
-
-        assert!(!solver.is_frozen(var![0]).unwrap());
     }
 }
