+++ glucose/src/simp.rs
@@ -352,92 +352,15 @@ impl Drop for Glucose {
 
 #[cfg(test)]
 mod test {
-    use std::{
-        sync::{Arc, Mutex},
-        thread,
-    };
-
     use super::Glucose;
     use rustsat::{
         lit,
-        solvers::{Solve, SolveStats, SolverResult},
+        solvers::{Solve, SolveStats},
         var,
     };
 
-    #[test]
-    fn build_destroy() {
-        let _solver = Glucose::default();
-    }
-
-    #[test]
-    fn build_two() {
-        let _solver1 = Glucose::default();
-        let _solver2 = Glucose::default();
-    }
-
-    #[test]
-    fn tiny_instance_sat() {
-        let mut solver = Glucose::default();
-        solver.add_binary(lit![0], !lit![1]).unwrap();
-        solver.add_binary(lit![1], !lit![2]).unwrap();
-        let ret = solver.solve();
-        match ret {
-            Err(e) => panic!("got error when solving: {}", e),
-            Ok(res) => assert_eq!(res, SolverResult::Sat),
-        }
-    }
-
-    #[test]
-    fn tiny_instance_unsat() {
-        let mut solver = Glucose::default();
-        solver.add_unit(!lit![0]).unwrap();
-        solver.add_binary(lit![0], !lit![1]).unwrap();
-        solver.add_binary(lit![1], !lit![2]).unwrap();
-        solver.add_unit(lit![2]).unwrap();
-        let ret = solver.solve();
-        match ret {
-            Err(e) => panic!("got error when solving: {}", e),
-            Ok(res) => assert_eq!(res, SolverResult::Unsat),
-        }
-    }
-
-    #[test]
-    fn tiny_instance_multithreaded_sat() {
-        let mutex_solver = Arc::new(Mutex::new(Glucose::default()));
-        let ret = {
-            let mut solver = mutex_solver.lock().unwrap();
-            solver.add_binary(lit![0], !lit![1]).unwrap();
-            solver.add_binary(lit![1], !lit![2]).unwrap();
-            solver.solve()
-        };
-        match ret {
-            Err(e) => panic!("got error when solving: {}", e),
-            Ok(res) => assert_eq!(res, SolverResult::Sat),
-        }
-
-        // Now in another thread
-        let s = mutex_solver.clone();
-        let ret = thread::spawn(move || {
-            let mut solver = s.lock().unwrap();
-            solver.solve()
-        })
-        .join()
-        .unwrap();
-        match ret {
-            Err(e) => panic!("got error when solving: {}", e),
-            Ok(res) => assert_eq!(res, SolverResult::Sat),
-        }
-
-        // Finally, back in the main thread
-        let ret = {
-            let mut solver = mutex_solver.lock().unwrap();
-            solver.solve()
-        };
-        match ret {
-            Err(e) => panic!("got error when solving: {}", e),
-            Ok(res) => assert_eq!(res, SolverResult::Sat),
-        }
-    }
+    rustsat_solvertests::basic_unittests!(Glucose);
+    rustsat_solvertests::freezing_unittests!(Glucose);
 
     #[test]
     fn backend_stats() {
