+++ minisat/src/core.rs
@@ -323,109 +323,14 @@ impl Drop for Minisat {
 
 #[cfg(test)]
 mod test {
-    use std::{
-        sync::{Arc, Mutex},
-        thread,
-    };
-
     use super::Minisat;
     use rustsat::{
         lit,
-        solvers::{Solve, SolveStats, SolverResult},
+        solvers::{Solve, SolveStats},
         var,
     };
 
-    #[test]
-    fn build_destroy() {
-        let _solver = Minisat::default();
-    }
-
-    #[test]
-    fn build_two() {
-        let _solver1 = Minisat::default();
-        let _solver2 = Minisat::default();
-    }
-
-    #[test]
-    fn tiny_instance_sat() {
-        let mut solver = Minisat::default();
-        solver.add_binary(lit![0], !lit![1]).unwrap();
-        solver.add_binary(lit![1], !lit![2]).unwrap();
-        let ret = solver.solve();
-        match ret {
-            Err(e) => panic!("got error when solving: {}", e),
-            Ok(res) => assert_eq!(res, SolverResult::Sat),
-        }
-    }
-
-    #[test]
-    fn tiny_instance_unsat() {
-        let mut solver = Minisat::default();
-        solver.add_unit(!lit![0]).unwrap();
-        solver.add_binary(lit![0], !lit![1]).unwrap();
-        solver.add_binary(lit![1], !lit![2]).unwrap();
-        solver.add_unit(lit![2]).unwrap();
-        let ret = solver.solve();
-        match ret {
-            Err(e) => panic!("got error when solving: {}", e),
-            Ok(res) => assert_eq!(res, SolverResult::Unsat),
-        }
-    }
-
-    #[test]
-    fn tiny_instance_multithreaded_sat() {
-        let mutex_solver = Arc::new(Mutex::new(Minisat::default()));
-        let ret = {
-            let mut solver = mutex_solver.lock().unwrap();
-            solver.add_binary(lit![0], !lit![1]).unwrap();
-            solver.add_binary(lit![1], !lit![2]).unwrap();
-            solver.solve()
-        };
-        match ret {
-            Err(e) => panic!("got error when solving: {}", e),
-            Ok(res) => assert_eq!(res, SolverResult::Sat),
-        }
-
-        // Now in another thread
-        let s = mutex_solver.clone();
-        let ret = thread::spawn(move || {
-            let mut solver = s.lock().unwrap();
-            solver.solve()
-        })
-        .join()
-        .unwrap();
-        match ret {
-            Err(e) => panic!("got error when solving: {}", e),
-            Ok(res) => assert_eq!(res, SolverResult::Sat),
-        }
-
-        // Now add to solver in other thread
-        let s = mutex_solver.clone();
-        let ret = thread::spawn(move || {
-            let mut solver = s.lock().unwrap();
-            solver.add_unit(!lit![0]).unwrap();
-            solver.add_unit(lit![2]).unwrap();
-            solver.solve()
-        })
-        .join()
-        .unwrap();
-        match ret {
-            Err(e) => panic!("got error when solving: {}", e),
-            Ok(res) => assert_eq!(res, SolverResult::Unsat),
-        }
-
-        // Finally, back in the main thread
-        let ret = {
-            let mut solver = mutex_solver.lock().unwrap();
-            solver.add_binary(lit![0], !lit![1]).unwrap();
-            solver.add_binary(lit![1], !lit![2]).unwrap();
-            solver.solve()
-        };
-        match ret {
-            Err(e) => panic!("got error when solving: {}", e),
-            Ok(res) => assert_eq!(res, SolverResult::Unsat),
-        }
-    }
+    rustsat_solvertests::basic_unittests!(Minisat);
 
     #[test]
     fn backend_stats() {
